@@ -13,7 +13,11 @@
         }
     ],
     "require": {
-        "php": ">=5.2"
+        "php": ">=7.2",
+        "ext-ctype": "*",
+        "ext-dom": "*",
+        "ext-libxml": "*",
+        "ext-iconv": "*"
     },
     "require-dev": {
         "simpletest/simpletest": "dev-master#72de02a7b80c6bb8864ef9bf66d41d2f58f826bd"
