@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Converts a stream of HTMLPurifier_Token into an HTMLPurifier_Node,
  * and back again.
@@ -9,10 +11,18 @@
  */
 class HTMLPurifier_Arborize
 {
-    public static function arborize($tokens, $config, $context) {
+    /**
+     * @param $tokens
+     * @param $config
+     *
+     * @return mixed
+     */
+    public static function arborize($tokens, $config)
+    {
         $definition = $config->getHTMLDefinition();
         $parent = new HTMLPurifier_Token_Start($definition->info_parent);
-        $stack = array($parent->toNode());
+        $stack = [$parent->toNode()];
+
         foreach ($tokens as $token) {
             $token->skip = null; // [MUT]
             $token->carryover = null; // [MUT]
@@ -26,25 +36,34 @@ class HTMLPurifier_Arborize
                 $r->endArmor = $token->armor;
                 continue;
             }
+
             $node = $token->toNode();
             $stack[count($stack)-1]->children[] = $node;
+
             if ($token instanceof HTMLPurifier_Token_Start) {
                 $stack[] = $node;
             }
         }
+
         //assert(count($stack) == 1);
         return $stack[0];
     }
 
-    public static function flatten($node, $config, $context) {
+    /**
+     * @param $node
+     *
+     * @return array
+     */
+    public static function flatten($node): array
+    {
         $level = 0;
-        $nodes = array($level => new HTMLPurifier_Queue(array($node)));
-        $closingTokens = array();
-        $tokens = array();
+        $nodes = [$level => new HTMLPurifier_Queue([$node])];
+        $closingTokens = [];
+        $tokens = [];
         do {
             while (!$nodes[$level]->isEmpty()) {
                 $node = $nodes[$level]->shift(); // FIFO
-                list($start, $end) = $node->toTokenPair();
+                [$start, $end] = $node->toTokenPair();
                 if ($level > 0) {
                     $tokens[] = $start;
                 }
@@ -66,6 +85,7 @@ class HTMLPurifier_Arborize
                 }
             }
         } while ($level > 0);
+
         return $tokens;
     }
 }
