@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Processes an entire attribute array for corrections needing multiple values.
  *
@@ -16,7 +18,6 @@
 
 abstract class HTMLPurifier_AttrTransform
 {
-
     /**
      * Abstract: makes changes to the attributes dependent on multiple values.
      *
@@ -34,9 +35,9 @@ abstract class HTMLPurifier_AttrTransform
      * @param array &$attr Attribute array to process (passed by reference)
      * @param string $css CSS to prepend
      */
-    public function prependCSS(&$attr, $css)
+    public function prependCSS(array &$attr, string $css)
     {
-        $attr['style'] = isset($attr['style']) ? $attr['style'] : '';
+        $attr['style'] = $attr['style'] ?? '';
         $attr['style'] = $css . $attr['style'];
     }
 
@@ -46,15 +47,15 @@ abstract class HTMLPurifier_AttrTransform
      * @param mixed $key Key of attribute to confiscate
      * @return mixed
      */
-    public function confiscateAttr(&$attr, $key)
+    public function confiscateAttr(array &$attr, $key)
     {
         if (!isset($attr[$key])) {
             return null;
         }
+
         $value = $attr[$key];
         unset($attr[$key]);
+
         return $value;
     }
 }
-
-// vim: et sw=4 sts=4
