@@ -1,45 +1,51 @@
 <?php
+declare(strict_types=1);
 
 /**
  * @todo Unit test
  */
 class HTMLPurifier_ContentSets
 {
-
     /**
      * List of content set strings (pipe separators) indexed by name.
+     *
      * @type array
      */
-    public $info = array();
+    public $info = [];
 
     /**
      * List of content set lookups (element => true) indexed by name.
+     *
      * @type array
      * @note This is in HTMLPurifier_HTMLDefinition->info_content_sets
      */
-    public $lookup = array();
+    public $lookup = [];
 
     /**
      * Synchronized list of defined content sets (keys of info).
+     *
      * @type array
      */
-    protected $keys = array();
+    protected $keys = [];
     /**
      * Synchronized list of defined content values (values of info).
+     *
      * @type array
      */
-    protected $values = array();
+    protected $values = [];
 
     /**
      * Merges in module's content sets, expands identifiers in the content
      * sets and populates the keys, values and lookup member variables.
+     *
      * @param HTMLPurifier_HTMLModule[] $modules List of HTMLPurifier_HTMLModule
      */
-    public function __construct($modules)
+    public function __construct(array $modules)
     {
         if (!is_array($modules)) {
-            $modules = array($modules);
+            $modules = [$modules];
         }
+
         // populate content_sets based on module hints
         // sorry, no way of overloading
         foreach ($modules as $module) {
@@ -53,11 +59,12 @@ class HTMLPurifier_ContentSets
                 }
             }
         }
+
         $old_lookup = false;
         while ($old_lookup !== $this->lookup) {
             $old_lookup = $this->lookup;
             foreach ($this->lookup as $i => $set) {
-                $add = array();
+                $add = [];
                 foreach ($set as $element => $x) {
                     if (isset($this->lookup[$element])) {
                         $add += $this->lookup[$element];
@@ -71,34 +78,43 @@ class HTMLPurifier_ContentSets
         foreach ($this->lookup as $key => $lookup) {
             $this->info[$key] = implode(' | ', array_keys($lookup));
         }
-        $this->keys   = array_keys($this->info);
+
+        $this->keys = array_keys($this->info);
         $this->values = array_values($this->info);
     }
 
     /**
      * Accepts a definition; generates and assigns a ChildDef for it
-     * @param HTMLPurifier_ElementDef $def HTMLPurifier_ElementDef reference
+     *
+     * @param HTMLPurifier_ElementDef $def    HTMLPurifier_ElementDef reference
      * @param HTMLPurifier_HTMLModule $module Module that defined the ElementDef
      */
-    public function generateChildDef(&$def, $module)
+    public function generateChildDef(HTMLPurifier_ElementDef &$def, HTMLPurifier_HTMLModule $module): void
     {
         if (!empty($def->child)) { // already done!
             return;
         }
+
         $content_model = $def->content_model;
         if (is_string($content_model)) {
             // Assume that $this->keys is alphanumeric
             $def->content_model = preg_replace_callback(
                 '/\b(' . implode('|', $this->keys) . ')\b/',
-                array($this, 'generateChildDefCallback'),
+                [$this, 'generateChildDefCallback'],
                 $content_model
             );
             //$def->content_model = str_replace(
             //    $this->keys, $this->values, $content_model);
         }
+
         $def->child = $this->getChildDef($def, $module);
     }
 
+    /**
+     * @param $matches
+     *
+     * @return mixed
+     */
     public function generateChildDefCallback($matches)
     {
         return $this->info[$matches[0]];
@@ -107,23 +123,28 @@ class HTMLPurifier_ContentSets
     /**
      * Instantiates a ChildDef based on content_model and content_model_type
      * member variables in HTMLPurifier_ElementDef
+     *
      * @note This will also defer to modules for custom HTMLPurifier_ChildDef
      *       subclasses that need content set expansion
-     * @param HTMLPurifier_ElementDef $def HTMLPurifier_ElementDef to have ChildDef extracted
+     *
+     * @param HTMLPurifier_ElementDef $def    HTMLPurifier_ElementDef to have ChildDef extracted
      * @param HTMLPurifier_HTMLModule $module Module that defined the ElementDef
+     *
      * @return HTMLPurifier_ChildDef corresponding to ElementDef
      */
-    public function getChildDef($def, $module)
+    public function getChildDef(HTMLPurifier_ElementDef $def, HTMLPurifier_HTMLModule $module): HTMLPurifier_ChildDef
     {
         $value = $def->content_model;
         if (is_object($value)) {
             trigger_error(
-                'Literal object child definitions should be stored in '.
+                'Literal object child definitions should be stored in ' .
                 'ElementDef->child not ElementDef->content_model',
                 E_USER_NOTICE
             );
+
             return $value;
         }
+
         switch ($def->content_model_type) {
             case 'required':
                 return new HTMLPurifier_ChildDef_Required($value);
@@ -134,37 +155,42 @@ class HTMLPurifier_ContentSets
             case 'custom':
                 return new HTMLPurifier_ChildDef_Custom($value);
         }
+
         // defer to its module
         $return = false;
         if ($module->defines_child_def) { // save a func call
             $return = $module->getChildDef($def);
         }
+
         if ($return !== false) {
             return $return;
         }
+
         // error-out
         trigger_error(
             'Could not determine which ChildDef class to instantiate',
             E_USER_ERROR
         );
+
         return false;
     }
 
     /**
      * Converts a string list of elements separated by pipes into
      * a lookup array.
+     *
      * @param string $string List of elements
+     *
      * @return array Lookup array of elements
      */
-    protected function convertToLookup($string)
+    protected function convertToLookup(string $string): array
     {
         $array = explode('|', str_replace(' ', '', $string));
-        $ret = array();
+        $ret = [];
         foreach ($array as $k) {
             $ret[$k] = true;
         }
+
         return $ret;
     }
 }
-
-// vim: et sw=4 sts=4
