@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Registry object that contains information about the current context.
  * @warning Is a bit buggy when variables are set to null: it thinks
@@ -9,19 +11,18 @@
  */
 class HTMLPurifier_Context
 {
-
     /**
      * Private array that stores the references.
      * @type array
      */
-    private $_storage = array();
+    private $_storage = [];
 
     /**
      * Registers a variable into the context.
      * @param string $name String name
      * @param mixed $ref Reference to variable to be registered
      */
-    public function register($name, &$ref)
+    public function register(string $name, &$ref)
     {
         if (array_key_exists($name, $this->_storage)) {
             trigger_error(
@@ -30,6 +31,7 @@ class HTMLPurifier_Context
             );
             return;
         }
+        
         $this->_storage[$name] =& $ref;
     }
 
@@ -39,7 +41,7 @@ class HTMLPurifier_Context
      * @param bool $ignore_error Boolean whether or not to ignore error
      * @return mixed
      */
-    public function &get($name, $ignore_error = false)
+    public function &get(string $name, bool $ignore_error = false)
     {
         if (!array_key_exists($name, $this->_storage)) {
             if (!$ignore_error) {
@@ -48,9 +50,11 @@ class HTMLPurifier_Context
                     E_USER_ERROR
                 );
             }
+
             $var = null; // so we can return by reference
             return $var;
         }
+
         return $this->_storage[$name];
     }
 
@@ -58,7 +62,7 @@ class HTMLPurifier_Context
      * Destroys a variable in the context.
      * @param string $name String name
      */
-    public function destroy($name)
+    public function destroy(string $name)
     {
         if (!array_key_exists($name, $this->_storage)) {
             trigger_error(
@@ -67,6 +71,7 @@ class HTMLPurifier_Context
             );
             return;
         }
+
         unset($this->_storage[$name]);
     }
 
@@ -75,7 +80,7 @@ class HTMLPurifier_Context
      * @param string $name String name
      * @return bool
      */
-    public function exists($name)
+    public function exists(string $name)
     {
         return array_key_exists($name, $this->_storage);
     }
@@ -84,12 +89,10 @@ class HTMLPurifier_Context
      * Loads a series of variables from an associative array
      * @param array $context_array Assoc array of variables to load
      */
-    public function loadArray($context_array)
+    public function loadArray(array $context_array)
     {
         foreach ($context_array as $key => $discard) {
             $this->register($key, $context_array[$key]);
         }
     }
 }
-
-// vim: et sw=4 sts=4
