@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Responsible for creating definition caches.
  */
@@ -8,17 +10,17 @@ class HTMLPurifier_DefinitionCacheFactory
     /**
      * @type array
      */
-    protected $caches = array('Serializer' => array());
+    protected $caches = ['Serializer' => []];
 
     /**
      * @type array
      */
-    protected $implementations = array();
+    protected $implementations = [];
 
     /**
      * @type HTMLPurifier_DefinitionCache_Decorator[]
      */
-    protected $decorators = array();
+    protected $decorators = [];
 
     /**
      * Initialize default decorators
@@ -30,67 +32,81 @@ class HTMLPurifier_DefinitionCacheFactory
 
     /**
      * Retrieves an instance of global definition cache factory.
+     *
      * @param HTMLPurifier_DefinitionCacheFactory $prototype
+     *
      * @return HTMLPurifier_DefinitionCacheFactory
      */
-    public static function instance($prototype = null)
+    public static function instance(HTMLPurifier_DefinitionCacheFactory $prototype = null)
     {
         static $instance;
         if ($prototype !== null) {
             $instance = $prototype;
         } elseif ($instance === null || $prototype === true) {
-            $instance = new HTMLPurifier_DefinitionCacheFactory();
+            $instance = new static();
             $instance->setup();
         }
+
         return $instance;
     }
 
     /**
      * Registers a new definition cache object
+     *
      * @param string $short Short name of cache object, for reference
-     * @param string $long Full class name of cache object, for construction
+     * @param string $long  Full class name of cache object, for construction
      */
-    public function register($short, $long)
+    public function register(string $short, string $long)
     {
         $this->implementations[$short] = $long;
     }
 
     /**
      * Factory method that creates a cache object based on configuration
-     * @param string $type Name of definitions handled by cache
+     *
+     * @param string              $type   Name of definitions handled by cache
      * @param HTMLPurifier_Config $config Config instance
+     *
      * @return mixed
+     * @throws HTMLPurifier_Exception
      */
-    public function create($type, $config)
+    public function create(string $type, HTMLPurifier_Config $config)
     {
         $method = $config->get('Cache.DefinitionImpl');
         if ($method === null) {
             return new HTMLPurifier_DefinitionCache_Null($type);
         }
+
         if (!empty($this->caches[$method][$type])) {
             return $this->caches[$method][$type];
         }
+
         if (isset($this->implementations[$method]) &&
             class_exists($class = $this->implementations[$method], false)) {
             $cache = new $class($type);
         } else {
-            if ($method != 'Serializer') {
+            if ($method !== 'Serializer') {
                 trigger_error("Unrecognized DefinitionCache $method, using Serializer instead", E_USER_WARNING);
             }
+
             $cache = new HTMLPurifier_DefinitionCache_Serializer($type);
         }
+
         foreach ($this->decorators as $decorator) {
             $new_cache = $decorator->decorate($cache);
             // prevent infinite recursion in PHP 4
             unset($cache);
             $cache = $new_cache;
         }
+
         $this->caches[$method][$type] = $cache;
+
         return $this->caches[$method][$type];
     }
 
     /**
      * Registers a decorator to add to all new cache objects
+     *
      * @param HTMLPurifier_DefinitionCache_Decorator|string $decorator An instance or the name of a decorator
      */
     public function addDecorator($decorator)
@@ -99,8 +115,7 @@ class HTMLPurifier_DefinitionCacheFactory
             $class = "HTMLPurifier_DefinitionCache_Decorator_$decorator";
             $decorator = new $class;
         }
+
         $this->decorators[$decorator->name] = $decorator;
     }
 }
-
-// vim: et sw=4 sts=4
