@@ -1,12 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * A UTF-8 specific character encoder that handles cleaning and transforming.
+ *
  * @note All functions in this class should be static.
  */
 class HTMLPurifier_Encoder
 {
-
     /**
      * Constructor throws fatal error if you attempt to instantiate class
      */
@@ -24,75 +26,85 @@ class HTMLPurifier_Encoder
 
     /**
      * iconv wrapper which mutes errors, but doesn't work around bugs.
-     * @param string $in Input encoding
-     * @param string $out Output encoding
+     *
+     * @param string $in   Input encoding
+     * @param string $out  Output encoding
      * @param string $text The text to convert
+     *
      * @return string
      */
     public static function unsafeIconv($in, $out, $text)
     {
-        set_error_handler(array('HTMLPurifier_Encoder', 'muteErrorHandler'));
+        set_error_handler(['HTMLPurifier_Encoder', 'muteErrorHandler']);
         $r = iconv($in, $out, $text);
         restore_error_handler();
+
         return $r;
     }
 
     /**
      * iconv wrapper which mutes errors and works around bugs.
-     * @param string $in Input encoding
-     * @param string $out Output encoding
+     *
+     * @param string $in   Input encoding
+     * @param string $out  Output encoding
      * @param string $text The text to convert
-     * @param int $max_chunk_size
+     * @param int    $max_chunk_size
+     *
      * @return string
      */
-    public static function iconv($in, $out, $text, $max_chunk_size = 8000)
+    public static function iconv(string $in, string $out, string $text, int $max_chunk_size = 8000)
     {
         $code = self::testIconvTruncateBug();
-        if ($code == self::ICONV_OK) {
+        if ($code === self::ICONV_OK) {
             return self::unsafeIconv($in, $out, $text);
-        } elseif ($code == self::ICONV_TRUNCATES) {
-            // we can only work around this if the input character set
-            // is utf-8
-            if ($in == 'utf-8') {
-                if ($max_chunk_size < 4) {
-                    trigger_error('max_chunk_size is too small', E_USER_WARNING);
-                    return false;
-                }
-                // split into 8000 byte chunks, but be careful to handle
-                // multibyte boundaries properly
-                if (($c = strlen($text)) <= $max_chunk_size) {
-                    return self::unsafeIconv($in, $out, $text);
-                }
-                $r = '';
-                $i = 0;
-                while (true) {
-                    if ($i + $max_chunk_size >= $c) {
-                        $r .= self::unsafeIconv($in, $out, substr($text, $i));
-                        break;
-                    }
-                    // wibble the boundary
-                    if (0x80 != (0xC0 & ord($text[$i + $max_chunk_size]))) {
-                        $chunk_size = $max_chunk_size;
-                    } elseif (0x80 != (0xC0 & ord($text[$i + $max_chunk_size - 1]))) {
-                        $chunk_size = $max_chunk_size - 1;
-                    } elseif (0x80 != (0xC0 & ord($text[$i + $max_chunk_size - 2]))) {
-                        $chunk_size = $max_chunk_size - 2;
-                    } elseif (0x80 != (0xC0 & ord($text[$i + $max_chunk_size - 3]))) {
-                        $chunk_size = $max_chunk_size - 3;
-                    } else {
-                        return false; // rather confusing UTF-8...
-                    }
-                    $chunk = substr($text, $i, $chunk_size); // substr doesn't mind overlong lengths
-                    $r .= self::unsafeIconv($in, $out, $chunk);
-                    $i += $chunk_size;
-                }
-                return $r;
-            } else {
+        }
+
+        // we can only work around this if the input character set
+        // is utf-8
+        if (($code === self::ICONV_TRUNCATES) && $in === 'utf-8') {
+            if ($max_chunk_size < 4) {
+                trigger_error('max_chunk_size is too small', E_USER_WARNING);
+
                 return false;
             }
-        } else {
-            return false;
+
+            // split into 8000 byte chunks, but be careful to handle
+            // multibyte boundaries properly
+            if (($c = strlen($text)) <= $max_chunk_size) {
+                return self::unsafeIconv($in, $out, $text);
+            }
+
+            $r = '';
+            $i = 0;
+
+            while (true) {
+                if ($i + $max_chunk_size >= $c) {
+                    $r .= self::unsafeIconv($in, $out, substr($text, $i));
+                    break;
+                }
+
+                // wibble the boundary
+                if ((0xC0 & ord($text[$i + $max_chunk_size])) !== 0x80) {
+                    $chunk_size = $max_chunk_size;
+                } elseif ((0xC0 & ord($text[$i + $max_chunk_size - 1])) !== 0x80) {
+                    $chunk_size = $max_chunk_size - 1;
+                } elseif ((0xC0 & ord($text[$i + $max_chunk_size - 2])) !== 0x80) {
+                    $chunk_size = $max_chunk_size - 2;
+                } elseif ((0xC0 & ord($text[$i + $max_chunk_size - 3])) !== 0x80) {
+                    $chunk_size = $max_chunk_size - 3;
+                } else {
+                    return false; // rather confusing UTF-8...
+                }
+
+                $chunk = substr($text, $i, $chunk_size); // substr doesn't mind overlong lengths
+                $r .= self::unsafeIconv($in, $out, $chunk);
+                $i += $chunk_size;
+            }
+
+            return $r;
         }
+
+        return false;
     }
 
     /**
@@ -110,7 +122,8 @@ class HTMLPurifier_Encoder
      * which simultaneously expand and restrict the set of allowed characters.
      *
      * @param string $str The string to clean
-     * @param bool $force_php
+     * @param bool   $force_php
+     *
      * @return string
      *
      * @note Just for reference, the non-SGML code points are 0 to 31 and
@@ -131,7 +144,7 @@ class HTMLPurifier_Encoder
      *       As of right now, only smart lossless character encoding converters
      *       would need that, and I'm probably not going to implement them.
      */
-    public static function cleanUTF8($str, $force_php = false)
+    public static function cleanUTF8(string $str, bool $force_php = false)
     {
         // UTF-8 validity is checked since PHP 4.3.5
         // This is an optimization: if the string is already valid UTF-8, no
@@ -144,8 +157,8 @@ class HTMLPurifier_Encoder
         }
 
         $mState = 0; // cached expected number of octets after the current octet
-                     // until the beginning of the next UTF8 character sequence
-        $mUcs4  = 0; // cached Unicode character
+        // until the beginning of the next UTF8 character sequence
+        $mUcs4 = 0; // cached Unicode character
         $mBytes = 1; // cached expected number of octets in the current sequence
 
         // original code involved an $out that was an array of Unicode
@@ -161,13 +174,13 @@ class HTMLPurifier_Encoder
         for ($i = 0; $i < $len; $i++) {
             $in = ord($str[$i]);
             $char .= $str[$i]; // append byte to char
-            if (0 == $mState) {
+            if ($mState === 0) {
                 // When mState is zero we expect either a US-ASCII character
                 // or a multi-octet sequence.
-                if (0 == (0x80 & ($in))) {
+                if ((0x80 & ($in)) === 0) {
                     // US-ASCII, pass straight through.
-                    if (($in <= 31 || $in == 127) &&
-                        !($in == 9 || $in == 13 || $in == 10) // save \r\t\n
+                    if (($in <= 31 || $in === 127) &&
+                        !($in === 9 || $in === 13 || $in === 10) // save \r\t\n
                     ) {
                         // control characters, remove
                     } else {
@@ -176,25 +189,25 @@ class HTMLPurifier_Encoder
                     // reset
                     $char = '';
                     $mBytes = 1;
-                } elseif (0xC0 == (0xE0 & ($in))) {
+                } elseif ((0xE0 & ($in)) === 0xC0) {
                     // First octet of 2 octet sequence
                     $mUcs4 = ($in);
                     $mUcs4 = ($mUcs4 & 0x1F) << 6;
                     $mState = 1;
                     $mBytes = 2;
-                } elseif (0xE0 == (0xF0 & ($in))) {
+                } elseif ((0xF0 & ($in)) === 0xE0) {
                     // First octet of 3 octet sequence
                     $mUcs4 = ($in);
                     $mUcs4 = ($mUcs4 & 0x0F) << 12;
                     $mState = 2;
                     $mBytes = 3;
-                } elseif (0xF0 == (0xF8 & ($in))) {
+                } elseif ((0xF8 & ($in)) === 0xF0) {
                     // First octet of 4 octet sequence
                     $mUcs4 = ($in);
                     $mUcs4 = ($mUcs4 & 0x07) << 18;
                     $mState = 3;
                     $mBytes = 4;
-                } elseif (0xF8 == (0xFC & ($in))) {
+                } elseif ((0xFC & ($in)) === 0xF8) {
                     // First octet of 5 octet sequence.
                     //
                     // This is illegal because the encoded codepoint must be
@@ -208,7 +221,7 @@ class HTMLPurifier_Encoder
                     $mUcs4 = ($mUcs4 & 0x03) << 24;
                     $mState = 4;
                     $mBytes = 5;
-                } elseif (0xFC == (0xFE & ($in))) {
+                } elseif ((0xFE & ($in)) === 0xFC) {
                     // First octet of 6 octet sequence, see comments for 5
                     // octet sequence.
                     $mUcs4 = ($in);
@@ -219,56 +232,54 @@ class HTMLPurifier_Encoder
                     // Current octet is neither in the US-ASCII range nor a
                     // legal first octet of a multi-octet sequence.
                     $mState = 0;
-                    $mUcs4  = 0;
+                    $mUcs4 = 0;
                     $mBytes = 1;
                     $char = '';
                 }
             } else {
-                // When mState is non-zero, we expect a continuation of the
-                // multi-octet sequence
-                if (0x80 == (0xC0 & ($in))) {
+                if ((0xC0 & ($in)) === 0x80) {
                     // Legal continuation.
                     $shift = ($mState - 1) * 6;
                     $tmp = $in;
                     $tmp = ($tmp & 0x0000003F) << $shift;
                     $mUcs4 |= $tmp;
 
-                    if (0 == --$mState) {
+                    if (--$mState === 0) {
                         // End of the multi-octet sequence. mUcs4 now contains
                         // the final Unicode codepoint to be output
 
                         // Check for illegal sequences and codepoints.
 
                         // From Unicode 3.1, non-shortest form is illegal
-                        if (((2 == $mBytes) && ($mUcs4 < 0x0080)) ||
-                            ((3 == $mBytes) && ($mUcs4 < 0x0800)) ||
-                            ((4 == $mBytes) && ($mUcs4 < 0x10000)) ||
+                        if ((($mBytes === 2) && ($mUcs4 < 0x0080)) ||
+                            (($mBytes === 3) && ($mUcs4 < 0x0800)) ||
+                            (($mBytes === 4) && ($mUcs4 < 0x10000)) ||
                             (4 < $mBytes) ||
                             // From Unicode 3.2, surrogate characters = illegal
-                            (($mUcs4 & 0xFFFFF800) == 0xD800) ||
+                            (($mUcs4 & 0xFFFFF800) === 0xD800) ||
                             // Codepoints outside the Unicode range are illegal
                             ($mUcs4 > 0x10FFFF)
                         ) {
 
-                        } elseif (0xFEFF != $mUcs4 && // omit BOM
-                            // check for valid Char unicode codepoints
-                            (
-                                0x9 == $mUcs4 ||
-                                0xA == $mUcs4 ||
-                                0xD == $mUcs4 ||
-                                (0x20 <= $mUcs4 && 0x7E >= $mUcs4) ||
-                                // 7F-9F is not strictly prohibited by XML,
-                                // but it is non-SGML, and thus we don't allow it
-                                (0xA0 <= $mUcs4 && 0xD7FF >= $mUcs4) ||
-                                (0xE000 <= $mUcs4 && 0xFFFD >= $mUcs4) ||
-                                (0x10000 <= $mUcs4 && 0x10FFFF >= $mUcs4)
-                            )
+                        } elseif ($mUcs4 !== 0xFEFF && // omit BOM
+                                  // check for valid Char unicode codepoints
+                                  (
+                                      $mUcs4 === 0x9 ||
+                                      $mUcs4 === 0xA ||
+                                      $mUcs4 === 0xD ||
+                                      (0x20 <= $mUcs4 && 0x7E >= $mUcs4) ||
+                                      // 7F-9F is not strictly prohibited by XML,
+                                      // but it is non-SGML, and thus we don't allow it
+                                      (0xA0 <= $mUcs4 && 0xD7FF >= $mUcs4) ||
+                                      (0xE000 <= $mUcs4 && 0xFFFD >= $mUcs4) ||
+                                      (0x10000 <= $mUcs4 && 0x10FFFF >= $mUcs4)
+                                  )
                         ) {
                             $out .= $char;
                         }
                         // initialize UTF8 cache (reset)
                         $mState = 0;
-                        $mUcs4  = 0;
+                        $mUcs4 = 0;
                         $mBytes = 1;
                         $char = '';
                     }
@@ -277,17 +288,19 @@ class HTMLPurifier_Encoder
                     // Incomplete multi-octet sequence.
                     // used to result in complete fail, but we'll reset
                     $mState = 0;
-                    $mUcs4  = 0;
+                    $mUcs4 = 0;
                     $mBytes = 1;
-                    $char ='';
+                    $char = '';
                 }
             }
         }
+
         return $out;
     }
 
     /**
      * Translates a Unicode codepoint into its corresponding UTF-8 character.
+     *
      * @note Based on Feyd's function at
      *       <http://forums.devnetwork.net/viewtopic.php?p=191404#191404>,
      *       which is in public domain.
@@ -297,6 +310,10 @@ class HTMLPurifier_Encoder
      * @note This is very similar to the unichr function in
      *       maintenance/generate-entity-file.php (although this is superior,
      *       due to its sanity checks).
+     *
+     * @param int $code
+     *
+     * @return string
      */
 
     // +----------+----------+----------+----------+
@@ -312,10 +329,10 @@ class HTMLPurifier_Encoder
     // | 00000000 | 00010000 | 11111111 | 11111111 | Defined upper limit of legal scalar codes
     // +----------+----------+----------+----------+
 
-    public static function unichr($code)
+    public static function unichr(int $code): string
     {
-        if ($code > 1114111 or $code < 0 or
-          ($code >= 55296 and $code <= 57343) ) {
+        if ($code > 1114111 || $code < 0 ||
+            ($code >= 55296 && $code <= 57343)) {
             // bits are set outside the "valid" range as defined
             // by UNICODE 4.1.0
             return '';
@@ -336,21 +353,25 @@ class HTMLPurifier_Encoder
                     $z = (($code >> 12) & 15) | 224;
                 } else {
                     $z = (($code >> 12) & 63) | 128;
-                    $w = (($code >> 18) & 7)  | 240;
+                    $w = (($code >> 18) & 7) | 240;
                 }
             }
         }
+
         // set up the actual character
         $ret = '';
         if ($w) {
             $ret .= chr($w);
         }
+
         if ($z) {
             $ret .= chr($z);
         }
+
         if ($y) {
             $ret .= chr($y);
         }
+
         $ret .= chr($x);
 
         return $ret;
@@ -359,51 +380,60 @@ class HTMLPurifier_Encoder
     /**
      * @return bool
      */
-    public static function iconvAvailable()
+    public static function iconvAvailable(): bool
     {
         static $iconv = null;
         if ($iconv === null) {
-            $iconv = function_exists('iconv') && self::testIconvTruncateBug() != self::ICONV_UNUSABLE;
+            $iconv = function_exists('iconv') && self::testIconvTruncateBug() !== self::ICONV_UNUSABLE;
         }
+
         return $iconv;
     }
 
     /**
      * Convert a string to UTF-8 based on configuration.
-     * @param string $str The string to convert
-     * @param HTMLPurifier_Config $config
+     *
+     * @param string               $str The string to convert
+     * @param HTMLPurifier_Config  $config
      * @param HTMLPurifier_Context $context
+     *
      * @return string
+     * @throws HTMLPurifier_Exception
      */
-    public static function convertToUTF8($str, $config, $context)
+    public static function convertToUTF8(string $str, HTMLPurifier_Config $config, HTMLPurifier_Context $context): string
     {
         $encoding = $config->get('Core.Encoding');
         if ($encoding === 'utf-8') {
             return $str;
         }
+
         static $iconv = null;
         if ($iconv === null) {
             $iconv = self::iconvAvailable();
         }
+
         if ($iconv && !$config->get('Test.ForceNoIconv')) {
             // unaffected by bugs, since UTF-8 support all characters
             $str = self::unsafeIconv($encoding, 'utf-8//IGNORE', $str);
             if ($str === false) {
                 // $encoding is not a valid encoding
                 trigger_error('Invalid encoding ' . $encoding, E_USER_ERROR);
+
                 return '';
             }
+
             // If the string is bjorked by Shift_JIS or a similar encoding
             // that doesn't support all of ASCII, convert the naughty
             // characters to their true byte-wise ASCII/UTF-8 equivalents.
-            $str = strtr($str, self::testEncodingSupportsASCII($encoding));
-            return $str;
-        } elseif ($encoding === 'iso-8859-1') {
-            $str = utf8_encode($str);
-            return $str;
+            return strtr($str, self::testEncodingSupportsASCII($encoding));
         }
-        $bug = HTMLPurifier_Encoder::testIconvTruncateBug();
-        if ($bug == self::ICONV_OK) {
+
+        if ($encoding === 'iso-8859-1') {
+            return utf8_encode($str);
+        }
+        $bug = static::testIconvTruncateBug();
+
+        if ($bug === static::ICONV_OK) {
             trigger_error('Encoding not supported, please install iconv', E_USER_ERROR);
         } else {
             trigger_error(
@@ -416,44 +446,53 @@ class HTMLPurifier_Encoder
 
     /**
      * Converts a string from UTF-8 based on configuration.
-     * @param string $str The string to convert
-     * @param HTMLPurifier_Config $config
+     *
+     * @param string               $str The string to convert
+     * @param HTMLPurifier_Config  $config
      * @param HTMLPurifier_Context $context
+     *
      * @return string
      * @note Currently, this is a lossy conversion, with unexpressable
      *       characters being omitted.
+     * @throws HTMLPurifier_Exception
      */
-    public static function convertFromUTF8($str, $config, $context)
+    public static function convertFromUTF8(?string $str, HTMLPurifier_Config $config, ?HTMLPurifier_Context $context): ?string
     {
         $encoding = $config->get('Core.Encoding');
         if ($escape = $config->get('Core.EscapeNonASCIICharacters')) {
-            $str = self::convertToASCIIDumbLossless($str);
+            $str = static::convertToASCIIDumbLossless($str);
         }
+
         if ($encoding === 'utf-8') {
             return $str;
         }
+
         static $iconv = null;
         if ($iconv === null) {
-            $iconv = self::iconvAvailable();
+            $iconv = static::iconvAvailable();
         }
+
         if ($iconv && !$config->get('Test.ForceNoIconv')) {
             // Undo our previous fix in convertToUTF8, otherwise iconv will barf
-            $ascii_fix = self::testEncodingSupportsASCII($encoding);
+            $ascii_fix = static::testEncodingSupportsASCII($encoding);
             if (!$escape && !empty($ascii_fix)) {
-                $clear_fix = array();
+                $clear_fix = [];
                 foreach ($ascii_fix as $utf8 => $native) {
                     $clear_fix[$utf8] = '';
                 }
                 $str = strtr($str, $clear_fix);
             }
+
             $str = strtr($str, array_flip($ascii_fix));
+
             // Normal stuff
-            $str = self::iconv('utf-8', $encoding . '//IGNORE', $str);
-            return $str;
-        } elseif ($encoding === 'iso-8859-1') {
-            $str = utf8_decode($str);
-            return $str;
+            return static::iconv('utf-8', $encoding . '//IGNORE', $str);
         }
+
+        if ($encoding === 'iso-8859-1') {
+            return utf8_decode($str);
+        }
+
         trigger_error('Encoding not supported', E_USER_ERROR);
         // You might be tempted to assume that the ASCII representation
         // might be OK, however, this is *not* universally true over all
@@ -463,21 +502,23 @@ class HTMLPurifier_Encoder
 
     /**
      * Lossless (character-wise) conversion of HTML to ASCII
+     *
      * @param string $str UTF-8 string to be converted to ASCII
+     *
      * @return string ASCII encoded string with non-ASCII character entity-ized
      * @warning Adapted from MediaWiki, claiming fair use: this is a common
      *       algorithm. If you disagree with this license fudgery,
      *       implement it yourself.
-     * @note Uses decimal numeric entities since they are best supported.
-     * @note This is a DUMB function: it has no concept of keeping
+     * @note    Uses decimal numeric entities since they are best supported.
+     * @note    This is a DUMB function: it has no concept of keeping
      *       character entities that the projected character encoding
      *       can allow. We could possibly implement a smart version
      *       but that would require it to also know which Unicode
      *       codepoints the charset supported (not an easy task).
-     * @note Sort of with cleanUTF8() but it assumes that $str is
+     * @note    Sort of with cleanUTF8() but it assumes that $str is
      *       well-formed UTF-8
      */
-    public static function convertToASCIIDumbLossless($str)
+    public static function convertToASCIIDumbLossless(string $str): string
     {
         $bytesleft = 0;
         $result = '';
@@ -489,11 +530,11 @@ class HTMLPurifier_Encoder
                 $result .= chr($bytevalue);
                 $bytesleft = 0;
             } elseif ($bytevalue <= 0xBF) { //10xx xxxx
-                $working = $working << 6;
+                $working <<= 6;
                 $working += ($bytevalue & 0x3F);
                 $bytesleft--;
                 if ($bytesleft <= 0) {
-                    $result .= "&#" . $working . ";";
+                    $result .= "&#${working};";
                 }
             } elseif ($bytevalue <= 0xDF) { //110x xxxx
                 $working = $bytevalue & 0x1F;
@@ -506,19 +547,20 @@ class HTMLPurifier_Encoder
                 $bytesleft = 3;
             }
         }
+
         return $result;
     }
 
     /** No bugs detected in iconv. */
-    const ICONV_OK = 0;
+    public const ICONV_OK = 0;
 
     /** Iconv truncates output if converting from UTF-8 to another
      *  character set with //IGNORE, and a non-encodable character is found */
-    const ICONV_TRUNCATES = 1;
+    public const ICONV_TRUNCATES = 1;
 
     /** Iconv does not support //IGNORE, making it unusable for
      *  transcoding purposes */
-    const ICONV_UNUSABLE = 2;
+    public const ICONV_UNUSABLE = 2;
 
     /**
      * glibc iconv has a known bug where it doesn't handle the magic
@@ -534,7 +576,7 @@ class HTMLPurifier_Encoder
      *
      * @return int Error code indicating severity of bug.
      */
-    public static function testIconvTruncateBug()
+    public static function testIconvTruncateBug(): int
     {
         static $code = null;
         if ($code === null) {
@@ -554,6 +596,7 @@ class HTMLPurifier_Encoder
                 $code = self::ICONV_OK;
             }
         }
+
         return $code;
     }
 
@@ -564,37 +607,41 @@ class HTMLPurifier_Encoder
      * encodings shouldn't ever fail.
      *
      * @param string $encoding Encoding name to test, as per iconv format
-     * @param bool $bypass Whether or not to bypass the precompiled arrays.
-     * @return Array of UTF-8 characters to their corresponding ASCII,
+     * @param bool   $bypass   Whether or not to bypass the precompiled arrays.
+     *
+     * @return array|bool of UTF-8 characters to their corresponding ASCII,
      *      which can be used to "undo" any overzealous iconv action.
      */
-    public static function testEncodingSupportsASCII($encoding, $bypass = false)
+    public static function testEncodingSupportsASCII(string $encoding, bool $bypass = false)
     {
         // All calls to iconv here are unsafe, proof by case analysis:
         // If ICONV_OK, no difference.
         // If ICONV_TRUNCATE, all calls involve one character inputs,
         // so bug is not triggered.
         // If ICONV_UNUSABLE, this call is irrelevant
-        static $encodings = array();
+        static $encodings = [];
         if (!$bypass) {
             if (isset($encodings[$encoding])) {
                 return $encodings[$encoding];
             }
+
             $lenc = strtolower($encoding);
             switch ($lenc) {
                 case 'shift_jis':
-                    return array("\xC2\xA5" => '\\', "\xE2\x80\xBE" => '~');
+                    return ["\xC2\xA5" => '\\', "\xE2\x80\xBE" => '~'];
                 case 'johab':
-                    return array("\xE2\x82\xA9" => '\\');
+                    return ["\xE2\x82\xA9" => '\\'];
             }
-            if (strpos($lenc, 'iso-8859-') === 0) {
-                return array();
+            if (strncmp($lenc, 'iso-8859-', 9) === 0) {
+                return [];
             }
         }
-        $ret = array();
+
+        $ret = [];
         if (self::unsafeIconv('UTF-8', $encoding, 'a') === false) {
             return false;
         }
+
         for ($i = 0x20; $i <= 0x7E; $i++) { // all printable ASCII chars
             $c = chr($i); // UTF-8 char
             $r = self::unsafeIconv('UTF-8', "$encoding//IGNORE", $c); // initial conversion
@@ -609,9 +656,9 @@ class HTMLPurifier_Encoder
                 $ret[self::unsafeIconv($encoding, 'UTF-8//IGNORE', $c)] = $c;
             }
         }
+
         $encodings[$encoding] = $ret;
+
         return $ret;
     }
 }
-
-// vim: et sw=4 sts=4
