@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Records errors for particular segments of an HTML document such as tokens,
  * attributes or CSS properties. They can contain error structs (which apply
@@ -8,14 +10,13 @@
  */
 class HTMLPurifier_ErrorStruct
 {
-
     /**
      * Possible values for $children first-key. Note that top-level structures
      * are automatically token-level.
      */
-    const TOKEN     = 0;
-    const ATTR      = 1;
-    const CSSPROP   = 2;
+    public const TOKEN = 0;
+    public const ATTR    = 1;
+    public const CSSPROP = 2;
 
     /**
      * Type of this struct.
@@ -37,7 +38,7 @@ class HTMLPurifier_ErrorStruct
      * Errors registered for this structure.
      * @type array
      */
-    public $errors = array();
+    public $errors = [];
 
     /**
      * Child ErrorStructs that are from this structure. For example, a TOKEN
@@ -45,19 +46,20 @@ class HTMLPurifier_ErrorStruct
      * array in structure: [TYPE]['identifier']
      * @type array
      */
-    public $children = array();
+    public $children = [];
 
     /**
-     * @param string $type
+     * @param string|int $type
      * @param string $id
      * @return mixed
      */
-    public function getChild($type, $id)
+    public function getChild($type, string$id)
     {
         if (!isset($this->children[$type][$id])) {
-            $this->children[$type][$id] = new HTMLPurifier_ErrorStruct();
+            $this->children[$type][$id] = new static();
             $this->children[$type][$id]->type = $type;
         }
+
         return $this->children[$type][$id];
     }
 
@@ -65,10 +67,8 @@ class HTMLPurifier_ErrorStruct
      * @param int $severity
      * @param string $message
      */
-    public function addError($severity, $message)
+    public function addError(int $severity, string $message)
     {
-        $this->errors[] = array($severity, $message);
+        $this->errors[] = [$severity, $message];
     }
 }
-
-// vim: et sw=4 sts=4
