@@ -1,7 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Generates HTML from tokens.
+ *
  * @todo Refactor interface so that configuration/context is determined
  *       upon instantiation, no need for messy generateFromTokens() calls
  * @todo Make some of the more internal functions protected, and have
@@ -9,15 +12,16 @@
  */
 class HTMLPurifier_Generator
 {
-
     /**
      * Whether or not generator should produce XML output.
+     *
      * @type bool
      */
     private $_xhtml = true;
 
     /**
      * :HACK: Whether or not generator should comment the insides of <script> tags.
+     *
      * @type bool
      */
     private $_scriptFix = false;
@@ -25,24 +29,28 @@ class HTMLPurifier_Generator
     /**
      * Cache of HTMLDefinition during HTML output to determine whether or
      * not attributes should be minimized.
+     *
      * @type HTMLPurifier_HTMLDefinition
      */
     private $_def;
 
     /**
      * Cache of %Output.SortAttr.
+     *
      * @type bool
      */
     private $_sortAttr;
 
     /**
      * Cache of %Output.FlashCompat.
+     *
      * @type bool
      */
     private $_flashCompat;
 
     /**
      * Cache of %Output.FixInnerHTML.
+     *
      * @type bool
      */
     private $_innerHTMLFix;
@@ -50,21 +58,25 @@ class HTMLPurifier_Generator
     /**
      * Stack for keeping track of object information when outputting IE
      * compatibility code.
+     *
      * @type array
      */
-    private $_flashStack = array();
+    private $_flashStack = [];
 
     /**
      * Configuration for the generator
+     *
      * @type HTMLPurifier_Config
      */
     protected $config;
 
     /**
-     * @param HTMLPurifier_Config $config
+     * @param HTMLPurifier_Config  $config
      * @param HTMLPurifier_Context $context
+     *
+     * @throws HTMLPurifier_Exception
      */
-    public function __construct($config, $context)
+    public function __construct(HTMLPurifier_Config $config, HTMLPurifier_Context $context)
     {
         $this->config = $config;
         $this->_scriptFix = $config->get('Output.CommentScriptContents');
@@ -77,10 +89,13 @@ class HTMLPurifier_Generator
 
     /**
      * Generates HTML from an array of tokens.
+     *
      * @param HTMLPurifier_Token[] $tokens Array of HTMLPurifier_Token
+     *
      * @return string Generated HTML
+     * @throws HTMLPurifier_Exception
      */
-    public function generateFromTokens($tokens)
+    public function generateFromTokens(array $tokens)
     {
         if (!$tokens) {
             return '';
@@ -90,7 +105,7 @@ class HTMLPurifier_Generator
         $html = '';
         for ($i = 0, $size = count($tokens); $i < $size; $i++) {
             if ($this->_scriptFix && $tokens[$i]->name === 'script'
-                && $i + 2 < $size && $tokens[$i+2] instanceof HTMLPurifier_Token_End) {
+                && $i + 2 < $size && $tokens[$i + 2] instanceof HTMLPurifier_Token_End) {
                 // script special case
                 // the contents of the script block must be ONE token
                 // for this to work.
@@ -105,17 +120,17 @@ class HTMLPurifier_Generator
             $tidy = new Tidy;
             $tidy->parseString(
                 $html,
-                array(
-                   'indent'=> true,
-                   'output-xhtml' => $this->_xhtml,
-                   'show-body-only' => true,
-                   'indent-spaces' => 2,
-                   'wrap' => 68,
-                ),
+                [
+                    'indent' => true,
+                    'output-xhtml' => $this->_xhtml,
+                    'show-body-only' => true,
+                    'indent-spaces' => 2,
+                    'wrap' => 68,
+                ],
                 'utf8'
             );
             $tidy->cleanRepair();
-            $html = (string) $tidy; // explicit cast necessary
+            $html = (string)$tidy; // explicit cast necessary
         }
 
         // Normalize newlines to system defined value
@@ -128,104 +143,123 @@ class HTMLPurifier_Generator
                 $html = str_replace("\n", $nl, $html);
             }
         }
+
         return $html;
     }
 
     /**
      * Generates HTML from a single token.
+     *
      * @param HTMLPurifier_Token $token HTMLPurifier_Token object.
+     *
      * @return string Generated HTML
      */
     public function generateFromToken($token)
     {
         if (!$token instanceof HTMLPurifier_Token) {
             trigger_error('Cannot generate HTML from non-HTMLPurifier_Token object', E_USER_WARNING);
-            return '';
 
-        } elseif ($token instanceof HTMLPurifier_Token_Start) {
+            return '';
+        }
+
+        if ($token instanceof HTMLPurifier_Token_Start) {
             $attr = $this->generateAttributes($token->attr, $token->name);
-            if ($this->_flashCompat) {
-                if ($token->name == "object") {
-                    $flash = new stdClass();
-                    $flash->attr = $token->attr;
-                    $flash->param = array();
-                    $this->_flashStack[] = $flash;
-                }
+            if ($this->_flashCompat && $token->name === "object") {
+                $flash = new stdClass();
+                $flash->attr = $token->attr;
+                $flash->param = [];
+                $this->_flashStack[] = $flash;
             }
+
             return '<' . $token->name . ($attr ? ' ' : '') . $attr . '>';
 
-        } elseif ($token instanceof HTMLPurifier_Token_End) {
-            $_extra = '';
-            if ($this->_flashCompat) {
-                if ($token->name == "object" && !empty($this->_flashStack)) {
-                    // doesn't do anything for now
-                }
-            }
-            return $_extra . '</' . $token->name . '>';
+        }
 
-        } elseif ($token instanceof HTMLPurifier_Token_Empty) {
-            if ($this->_flashCompat && $token->name == "param" && !empty($this->_flashStack)) {
-                $this->_flashStack[count($this->_flashStack)-1]->param[$token->attr['name']] = $token->attr['value'];
+        if ($token instanceof HTMLPurifier_Token_End) {
+            $_extra = '';
+            if ($this->_flashCompat && $token->name === 'object' && !empty($this->_flashStack)) {
+                // doesn't do anything for now
+            }
+
+            return $_extra . '</' . $token->name . '>';
+        }
+
+        if ($token instanceof HTMLPurifier_Token_Empty) {
+            if ($this->_flashCompat && $token->name === 'param' && !empty($this->_flashStack)) {
+                $this->_flashStack[count($this->_flashStack) - 1]->param[$token->attr['name']] = $token->attr['value'];
             }
             $attr = $this->generateAttributes($token->attr, $token->name);
-             return '<' . $token->name . ($attr ? ' ' : '') . $attr .
-                ( $this->_xhtml ? ' /': '' ) // <br /> v. <br>
-                . '>';
 
-        } elseif ($token instanceof HTMLPurifier_Token_Text) {
-            return $this->escape($token->data, ENT_NOQUOTES);
-
-        } elseif ($token instanceof HTMLPurifier_Token_Comment) {
-            return '<!--' . $token->data . '-->';
-        } else {
-            return '';
+            return '<' . $token->name . ($attr ? ' ' : '') . $attr .
+                   ($this->_xhtml ? ' /' : '') // <br /> v. <br>
+                   . '>';
 
         }
+
+        if ($token instanceof HTMLPurifier_Token_Text) {
+            return $this->escape($token->data, ENT_NOQUOTES);
+        }
+
+        if ($token instanceof HTMLPurifier_Token_Comment) {
+            return '<!--' . $token->data . '-->';
+        }
+
+        return '';
     }
 
     /**
      * Special case processor for the contents of script tags
+     *
      * @param HTMLPurifier_Token $token HTMLPurifier_Token object.
+     *
      * @return string
      * @warning This runs into problems if there's already a literal
      *          --> somewhere inside the script contents.
      */
-    public function generateScriptFromToken($token)
+    public function generateScriptFromToken(HTMLPurifier_Token $token)
     {
         if (!$token instanceof HTMLPurifier_Token_Text) {
             return $this->generateFromToken($token);
         }
+
         // Thanks <http://lachy.id.au/log/2005/05/script-comments>
         $data = preg_replace('#//\s*$#', '', $token->data);
+
         return '<!--//--><![CDATA[//><!--' . "\n" . trim($data) . "\n" . '//--><!]]>';
     }
 
     /**
      * Generates attribute declarations from attribute array.
+     *
      * @note This does not include the leading or trailing space.
-     * @param array $assoc_array_of_attributes Attribute array
-     * @param string $element Name of element attributes are for, used to check
-     *        attribute minimization.
+     *
+     * @param array  $assoc_array_of_attributes Attribute array
+     * @param string $element                   Name of element attributes are for, used to check
+     *                                          attribute minimization.
+     *
      * @return string Generated HTML fragment for insertion.
      */
-    public function generateAttributes($assoc_array_of_attributes, $element = '')
+    public function generateAttributes(array $assoc_array_of_attributes, string $element = '')
     {
         $html = '';
         if ($this->_sortAttr) {
             ksort($assoc_array_of_attributes);
         }
+
         foreach ($assoc_array_of_attributes as $key => $value) {
             if (!$this->_xhtml) {
                 // Remove namespaced attributes
                 if (strpos($key, ':') !== false) {
                     continue;
                 }
+
                 // Check if we should minimize the attribute: val="val" -> val
                 if ($element && !empty($this->_def->info[$element]->attr[$key]->minimized)) {
                     $html .= $key . ' ';
                     continue;
                 }
             }
+
             // Workaround for Internet Explorer innerHTML bug.
             // Essentially, Internet Explorer, when calculating
             // innerHTML, omits quotes if there are no instances of
@@ -257,30 +291,33 @@ class HTMLPurifier_Generator
                     }
                 }
             }
-            $html .= $key.'="'.$this->escape($value).'" ';
+
+            $html .= $key . '="' . $this->escape($value) . '" ';
         }
+
         return rtrim($html);
     }
 
     /**
      * Escapes raw text data.
+     *
+     * @param string $string String data to escape for HTML.
+     * @param int    $quote  Quoting style, like htmlspecialchars. ENT_NOQUOTES is
+     *                       permissible for non-attribute output.
+     *
+     * @return string escaped data.
      * @todo This really ought to be protected, but until we have a facility
      *       for properly generating HTML here w/o using tokens, it stays
      *       public.
-     * @param string $string String data to escape for HTML.
-     * @param int $quote Quoting style, like htmlspecialchars. ENT_NOQUOTES is
-     *               permissible for non-attribute output.
-     * @return string escaped data.
      */
-    public function escape($string, $quote = null)
+    public function escape(string $string, int $quote = null)
     {
         // Workaround for APC bug on Mac Leopard reported by sidepodcast
         // http://htmlpurifier.org/phorum/read.php?3,4823,4846
         if ($quote === null) {
             $quote = ENT_COMPAT;
         }
+
         return htmlspecialchars($string, $quote, 'UTF-8');
     }
 }
-
-// vim: et sw=4 sts=4
