@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Definition of the purified HTML that describes allowed children,
  * attributes, and many other things.
@@ -17,7 +19,7 @@
  * Purifier internals. Many of them, however, are public, and may be
  * edited by userspace code to tweak the behavior of HTMLDefinition.
  *
- * @note This class is inspected by Printer_HTMLDefinition; please
+ * @note    This class is inspected by Printer_HTMLDefinition; please
  *       update that class if things here change.
  *
  * @warning Directives that change this object's structure must be in
@@ -25,23 +27,25 @@
  */
 class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
 {
-
     // FULLY-PUBLIC VARIABLES ---------------------------------------------
 
     /**
      * Associative array of element names to HTMLPurifier_ElementDef.
+     *
      * @type HTMLPurifier_ElementDef[]
      */
-    public $info = array();
+    public $info = [];
 
     /**
      * Associative array of global attribute name to attribute definition.
+     *
      * @type array
      */
-    public $info_global_attr = array();
+    public $info_global_attr = [];
 
     /**
      * String name of parent element HTML will be going into.
+     *
      * @type string
      */
     public $info_parent = 'div';
@@ -49,12 +53,14 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
     /**
      * Definition for parent element, allows parent element to be a
      * tag that's not allowed inside the HTML fragment.
+     *
      * @type HTMLPurifier_ElementDef
      */
     public $info_parent_def;
 
     /**
      * String name of element used to wrap inline elements in block context.
+     *
      * @type string
      * @note This is rarely used except for BLOCKQUOTEs in strict mode
      */
@@ -62,55 +68,61 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
 
     /**
      * Associative array of deprecated tag name to HTMLPurifier_TagTransform.
+     *
      * @type array
      */
-    public $info_tag_transform = array();
+    public $info_tag_transform = [];
 
     /**
      * Indexed list of HTMLPurifier_AttrTransform to be performed before validation.
+     *
      * @type HTMLPurifier_AttrTransform[]
      */
-    public $info_attr_transform_pre = array();
+    public $info_attr_transform_pre = [];
 
     /**
      * Indexed list of HTMLPurifier_AttrTransform to be performed after validation.
+     *
      * @type HTMLPurifier_AttrTransform[]
      */
-    public $info_attr_transform_post = array();
+    public $info_attr_transform_post = [];
 
     /**
      * Nested lookup array of content set name (Block, Inline) to
      * element name to whether or not it belongs in that content set.
+     *
      * @type array
      */
-    public $info_content_sets = array();
+    public $info_content_sets = [];
 
     /**
      * Indexed list of HTMLPurifier_Injector to be used.
+     *
      * @type HTMLPurifier_Injector[]
      */
-    public $info_injector = array();
+    public $info_injector = [];
 
     /**
      * Doctype object
+     *
      * @type HTMLPurifier_Doctype
      */
     public $doctype;
-
-
 
     // RAW CUSTOMIZATION STUFF --------------------------------------------
 
     /**
      * Adds a custom attribute to a pre-existing element
+     *
      * @note This is strictly convenience, and does not have a corresponding
      *       method in HTMLPurifier_HTMLModule
+     *
      * @param string $element_name Element name to add attribute to
-     * @param string $attr_name Name of attribute
-     * @param mixed $def Attribute definition, can be string or object, see
-     *             HTMLPurifier_AttrTypes for details
+     * @param string $attr_name    Name of attribute
+     * @param mixed  $def          Attribute definition, can be string or object, see
+     *                             HTMLPurifier_AttrTypes for details
      */
-    public function addAttribute($element_name, $attr_name, $def)
+    public function addAttribute(string $element_name, string $attr_name, $def)
     {
         $module = $this->getAnonymousModule();
         if (!isset($module->info[$element_name])) {
@@ -118,42 +130,47 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
         } else {
             $element = $module->info[$element_name];
         }
+
         $element->attr[$attr_name] = $def;
     }
 
     /**
      * Adds a custom element to your HTML definition
+     *
      * @see HTMLPurifier_HTMLModule::addElement() for detailed
      *       parameter and return value descriptions.
      */
-    public function addElement($element_name, $type, $contents, $attr_collections, $attributes = array())
+    public function addElement(string $element_name, $type, $contents, $attr_collections, array $attributes = [])
     {
         $module = $this->getAnonymousModule();
+
         // assume that if the user is calling this, the element
         // is safe. This may not be a good idea
-        $element = $module->addElement($element_name, $type, $contents, $attr_collections, $attributes);
-        return $element;
+        return $module->addElement($element_name, $type, $contents, $attr_collections, $attributes);
     }
 
     /**
      * Adds a blank element to your HTML definition, for overriding
      * existing behavior
+     *
      * @param string $element_name
+     *
      * @return HTMLPurifier_ElementDef
      * @see HTMLPurifier_HTMLModule::addBlankElement() for detailed
      *       parameter and return value descriptions.
      */
-    public function addBlankElement($element_name)
+    public function addBlankElement(string $element_name)
     {
-        $module  = $this->getAnonymousModule();
-        $element = $module->addBlankElement($element_name);
-        return $element;
+        $module = $this->getAnonymousModule();
+
+        return $module->addBlankElement($element_name);
     }
 
     /**
      * Retrieves a reference to the anonymous module, so you can
      * bust out advanced features without having to make your own
      * module.
+     *
      * @return HTMLPurifier_HTMLModule
      */
     public function getAnonymousModule()
@@ -162,6 +179,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
             $this->_anonModule = new HTMLPurifier_HTMLModule();
             $this->_anonModule->name = 'Anonymous';
         }
+
         return $this->_anonModule;
     }
 
@@ -190,7 +208,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
     /**
      * @param HTMLPurifier_Config $config
      */
-    protected function doSetup($config)
+    protected function doSetup(HTMLPurifier_Config $config)
     {
         $this->processModules($config);
         $this->setupConfigStuff($config);
@@ -198,16 +216,19 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
 
         // cleanup some of the element definitions
         foreach ($this->info as $k => $v) {
-            unset($this->info[$k]->content_model);
-            unset($this->info[$k]->content_model_type);
+            unset(
+                $this->info[$k]->content_model,
+                $this->info[$k]->content_model_type
+            );
         }
     }
 
     /**
      * Extract out the information from the manager
+     *
      * @param HTMLPurifier_Config $config
      */
-    protected function processModules($config)
+    protected function processModules(HTMLPurifier_Config $config)
     {
         if ($this->_anonModule) {
             // for user specific changes
@@ -250,15 +271,19 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                 }
             }
         }
+
         $this->info = $this->manager->getElements();
         $this->info_content_sets = $this->manager->contentSets->lookup;
     }
 
     /**
      * Sets up stuff based on config. We need a better way of doing this.
+     *
      * @param HTMLPurifier_Config $config
+     *
+     * @throws HTMLPurifier_Exception
      */
-    protected function setupConfigStuff($config)
+    protected function setupConfigStuff(HTMLPurifier_Config $config)
     {
         $block_wrapper = $config->get('HTML.BlockWrapper');
         if (isset($this->info_content_sets['Block'][$block_wrapper])) {
@@ -284,7 +309,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
         }
 
         // support template text
-        $support = "(for information on implementing this, see the support forums) ";
+        $support = '(for information on implementing this, see the support forums) ';
 
         // setup allowed elements -----------------------------------------
 
@@ -294,7 +319,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
         if (!is_array($allowed_elements) && !is_array($allowed_attributes)) {
             $allowed = $config->get('HTML.Allowed');
             if (is_string($allowed)) {
-                list($allowed_elements, $allowed_attributes) = $this->parseTinyMCEAllowedList($allowed);
+                [$allowed_elements, $allowed_attributes] = $this->parseTinyMCEAllowedList($allowed);
             }
         }
 
@@ -303,6 +328,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                 if (!isset($allowed_elements[$name])) {
                     unset($this->info[$name]);
                 }
+
                 unset($allowed_elements[$name]);
             }
             // emit errors
@@ -320,16 +346,18 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
             // global attributes. It's possible that userland code uses
             // it, but HTMLModuleManager doesn't!
             foreach ($this->info_global_attr as $attr => $x) {
-                $keys = array($attr, "*@$attr", "*.$attr");
+                $keys = [$attr, "*@$attr", "*.$attr"];
                 $delete = true;
                 foreach ($keys as $key) {
                     if ($delete && isset($allowed_attributes[$key])) {
                         $delete = false;
                     }
+
                     if (isset($allowed_attributes_mutable[$key])) {
                         unset($allowed_attributes_mutable[$key]);
                     }
                 }
+
                 if ($delete) {
                     unset($this->info_global_attr[$attr]);
                 }
@@ -337,7 +365,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
 
             foreach ($this->info as $tag => $info) {
                 foreach ($info->attr as $attr => $x) {
-                    $keys = array("$tag@$attr", $attr, "*@$attr", "$tag.$attr", "*.$attr");
+                    $keys = ["$tag@$attr", $attr, "*@$attr", "$tag.$attr", "*.$attr"];
                     $delete = true;
                     foreach ($keys as $key) {
                         if ($delete && isset($allowed_attributes[$key])) {
@@ -347,6 +375,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                             unset($allowed_attributes_mutable[$key]);
                         }
                     }
+
                     if ($delete) {
                         if ($this->info[$tag]->attr[$attr]->required) {
                             trigger_error(
@@ -355,6 +384,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                                 E_USER_WARNING
                             );
                         }
+
                         unset($this->info[$tag]->attr[$attr]);
                     }
                 }
@@ -381,11 +411,11 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                             }
                             break;
                         }
-                        // otherwise fall through
+                    // otherwise fall through
                     case 1:
                         $attribute = htmlspecialchars($bits[0]);
                         trigger_error(
-                            "Global attribute '$attribute' is not ".
+                            "Global attribute '$attribute' is not " .
                             "supported in any elements $support",
                             E_USER_WARNING
                         );
@@ -395,8 +425,7 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
         }
 
         // setup forbidden elements ---------------------------------------
-
-        $forbidden_elements   = $config->get('HTML.ForbiddenElements');
+        $forbidden_elements = $config->get('HTML.ForbiddenElements');
         $forbidden_attributes = $config->get('HTML.ForbiddenAttributes');
 
         foreach ($this->info as $tag => $info) {
@@ -407,28 +436,30 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
             foreach ($info->attr as $attr => $x) {
                 if (isset($forbidden_attributes["$tag@$attr"]) ||
                     isset($forbidden_attributes["*@$attr"]) ||
-                    isset($forbidden_attributes[$attr])
-                ) {
+                    isset($forbidden_attributes[$attr])) {
                     unset($this->info[$tag]->attr[$attr]);
                     continue;
-                } elseif (isset($forbidden_attributes["$tag.$attr"])) { // this segment might get removed eventually
+                }
+
+                if (isset($forbidden_attributes["$tag.$attr"])) { // this segment might get removed eventually
                     // $tag.$attr are not user supplied, so no worries!
                     trigger_error(
                         "Error with $tag.$attr: tag.attr syntax not supported for " .
-                        "HTML.ForbiddenAttributes; use tag@attr instead",
+                        'HTML.ForbiddenAttributes; use tag@attr instead',
                         E_USER_WARNING
                     );
                 }
             }
         }
+
         foreach ($forbidden_attributes as $key => $v) {
             if (strlen($key) < 2) {
                 continue;
             }
-            if ($key[0] != '*') {
+            if ($key[0] !== '*') {
                 continue;
             }
-            if ($key[1] == '.') {
+            if ($key[1] === '.') {
                 trigger_error(
                     "Error with $key: *.attr syntax not supported for HTML.ForbiddenAttributes; use attr instead",
                     E_USER_WARNING
@@ -449,18 +480,21 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
     /**
      * Parses a TinyMCE-flavored Allowed Elements and Attributes list into
      * separate lists for processing. Format is element[attr1|attr2],element2...
+     *
      * @warning Although it's largely drawn from TinyMCE's implementation,
      *      it is different, and you'll probably have to modify your lists
+     *
      * @param array $list String list to parse
+     *
      * @return array
-     * @todo Give this its own class, probably static interface
+     * @todo    Give this its own class, probably static interface
      */
     public function parseTinyMCEAllowedList($list)
     {
-        $list = str_replace(array(' ', "\t"), '', $list);
+        $list = str_replace([' ', "\t"], '', $list);
 
-        $elements = array();
-        $attributes = array();
+        $elements = [];
+        $attributes = [];
 
         $chunks = preg_split('/(,|[\n\r]+)/', $list);
         foreach ($chunks as $chunk) {
@@ -472,22 +506,24 @@ class HTMLPurifier_HTMLDefinition extends HTMLPurifier_Definition
                 $element = $chunk;
                 $attr = false;
             } else {
-                list($element, $attr) = explode('[', $chunk);
+                [$element, $attr] = explode('[', $chunk);
             }
+
             if ($element !== '*') {
                 $elements[$element] = true;
             }
+
             if (!$attr) {
                 continue;
             }
-            $attr = substr($attr, 0, strlen($attr) - 1); // remove trailing ]
+
+            $attr = substr($attr, 0, -1); // remove trailing ]
             $attr = explode('|', $attr);
             foreach ($attr as $key) {
                 $attributes["$element.$key"] = true;
             }
         }
-        return array($elements, $attributes);
+
+        return [$elements, $attributes];
     }
 }
-
-// vim: et sw=4 sts=4
