@@ -1,18 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Class responsible for generating HTMLPurifier_Language objects, managing
  * caching and fallbacks.
+ *
  * @note Thanks to MediaWiki for the general logic, although this version
  *       has been entirely rewritten
  * @todo Serialized cache for languages
  */
 class HTMLPurifier_LanguageFactory
 {
-
     /**
      * Cache of language code information used to load HTMLPurifier_Language objects.
      * Structure is: $factory->cache[$language_code][$key] = $value
+     *
      * @type array
      */
     public $cache;
@@ -20,12 +23,14 @@ class HTMLPurifier_LanguageFactory
     /**
      * Valid keys in the HTMLPurifier_Language object. Designates which
      * variables to slurp out of a message file.
+     *
      * @type array
      */
-    public $keys = array('fallback', 'messages', 'errorNames');
+    public $keys = ['fallback', 'messages', 'errorNames'];
 
     /**
      * Instance to validate language codes.
+     *
      * @type HTMLPurifier_AttrDef_Lang
      *
      */
@@ -34,42 +39,49 @@ class HTMLPurifier_LanguageFactory
     /**
      * Cached copy of dirname(__FILE__), directory of current file without
      * trailing slash.
+     *
      * @type string
      */
     protected $dir;
 
     /**
      * Keys whose contents are a hash map and can be merged.
+     *
      * @type array
      */
-    protected $mergeable_keys_map = array('messages' => true, 'errorNames' => true);
+    protected $mergeable_keys_map = ['messages' => true, 'errorNames' => true];
 
     /**
      * Keys whose contents are a list and can be merged.
+     *
      * @value array lookup
      */
-    protected $mergeable_keys_list = array();
+    protected $mergeable_keys_list = [];
 
     /**
      * Retrieve sole instance of the factory.
+     *
      * @param HTMLPurifier_LanguageFactory $prototype Optional prototype to overload sole instance with,
-     *                   or bool true to reset to default factory.
+     *                                                or bool true to reset to default factory.
+     *
      * @return HTMLPurifier_LanguageFactory
      */
-    public static function instance($prototype = null)
+    public static function instance(HTMLPurifier_LanguageFactory $prototype = null)
     {
         static $instance = null;
         if ($prototype !== null) {
             $instance = $prototype;
-        } elseif ($instance === null || $prototype == true) {
-            $instance = new HTMLPurifier_LanguageFactory();
+        } elseif ($instance === null || $prototype === true) {
+            $instance = new static();
             $instance->setup();
         }
+
         return $instance;
     }
 
     /**
      * Sets up the singleton, much like a constructor
+     *
      * @note Prevents people from getting this outside of the singleton
      */
     public function setup()
@@ -80,12 +92,15 @@ class HTMLPurifier_LanguageFactory
 
     /**
      * Creates a language object, handles class fallbacks
-     * @param HTMLPurifier_Config $config
+     *
+     * @param HTMLPurifier_Config  $config
      * @param HTMLPurifier_Context $context
-     * @param bool|string $code Code to override configuration with. Private parameter.
+     * @param bool|string          $code Code to override configuration with. Private parameter.
+     *
      * @return HTMLPurifier_Language
+     * @throws HTMLPurifier_Exception
      */
-    public function create($config, $context, $code = false)
+    public function create(HTMLPurifier_Config $config, HTMLPurifier_Context $context, $code = false)
     {
         // validate language code
         if ($code === false) {
@@ -97,6 +112,7 @@ class HTMLPurifier_LanguageFactory
         } else {
             $code = $this->validator->validate($code, $config, $context);
         }
+
         if ($code === false) {
             $code = 'en'; // malformed code becomes English
         }
@@ -104,48 +120,58 @@ class HTMLPurifier_LanguageFactory
         $pcode = str_replace('-', '_', $code); // make valid PHP classname
         static $depth = 0; // recursion protection
 
-        if ($code == 'en') {
+        if ($code === 'en') {
             $lang = new HTMLPurifier_Language($config, $context);
         } else {
             $class = 'HTMLPurifier_Language_' . $pcode;
-            $file  = $this->dir . '/Language/classes/' . $code . '.php';
+            $file = $this->dir . '/Language/classes/' . $code . '.php';
+
             if (file_exists($file) || class_exists($class, false)) {
                 $lang = new $class($config, $context);
             } else {
                 // Go fallback
                 $raw_fallback = $this->getFallbackFor($code);
-                $fallback = $raw_fallback ? $raw_fallback : 'en';
+                $fallback = $raw_fallback ?: 'en';
                 $depth++;
                 $lang = $this->create($config, $context, $fallback);
+
                 if (!$raw_fallback) {
                     $lang->error = true;
                 }
+
                 $depth--;
             }
         }
+
         $lang->code = $code;
+
         return $lang;
     }
 
     /**
      * Returns the fallback language for language
+     *
      * @note Loads the original language into cache
+     *
      * @param string $code language code
+     *
      * @return string|bool
      */
-    public function getFallbackFor($code)
+    public function getFallbackFor(string $code)
     {
         $this->loadLanguage($code);
+
         return $this->cache[$code]['fallback'];
     }
 
     /**
      * Loads language into the cache, handles message file and fallbacks
+     *
      * @param string $code language code
      */
-    public function loadLanguage($code)
+    public function loadLanguage(string $code)
     {
-        static $languages_seen = array(); // recursion guard
+        static $languages_seen = []; // recursion guard
 
         // abort if we've already loaded it
         if (isset($this->cache[$code])) {
@@ -153,16 +179,16 @@ class HTMLPurifier_LanguageFactory
         }
 
         // generate filename
-        $filename = $this->dir . '/Language/messages/' . $code . '.php';
+        $filename = "{$this->dir}/Language/messages/${code}.php";
 
         // default fallback : may be overwritten by the ensuing include
-        $fallback = ($code != 'en') ? 'en' : false;
+        $fallback = ($code !== 'en') ? 'en' : false;
 
         // load primary localisation
         if (!file_exists($filename)) {
             // skip the include: will rely solely on fallback
             $filename = $this->dir . '/Language/messages/en.php';
-            $cache = array();
+            $cache = [];
         } else {
             include $filename;
             $cache = compact($this->keys);
@@ -178,8 +204,10 @@ class HTMLPurifier_LanguageFactory
                     $code,
                     E_USER_ERROR
                 );
+
                 $fallback = 'en';
             }
+
             $language_seen[$code] = true;
 
             // load the fallback recursively
@@ -188,9 +216,9 @@ class HTMLPurifier_LanguageFactory
 
             // merge fallback with current language
             foreach ($this->keys as $key) {
-                if (isset($cache[$key]) && isset($fallback_cache[$key])) {
+                if (isset($cache[$key], $fallback_cache[$key])) {
                     if (isset($this->mergeable_keys_map[$key])) {
-                        $cache[$key] = $cache[$key] + $fallback_cache[$key];
+                        $cache[$key] += $fallback_cache[$key];
                     } elseif (isset($this->mergeable_keys_list[$key])) {
                         $cache[$key] = array_merge($fallback_cache[$key], $cache[$key]);
                     }
@@ -202,8 +230,7 @@ class HTMLPurifier_LanguageFactory
 
         // save to cache for later retrieval
         $this->cache[$code] = $cache;
+
         return;
     }
 }
-
-// vim: et sw=4 sts=4
