@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 /**
  * Abstract base node class that all others inherit from.
@@ -27,7 +28,7 @@ abstract class HTMLPurifier_Node
      * Currently, valid values are "ValidateAttributes".
      * @type array
      */
-    public $armor = array();
+    public $armor = [];
 
     /**
      * When true, this node should be ignored as non-existent.
