@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 if (!defined('HTMLPURIFIER_PREFIX')) {
-    define('HTMLPURIFIER_PREFIX', dirname(__FILE__));
+    define('HTMLPURIFIER_PREFIX', __DIR__);
 }
