@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @file
  * Defines a function wrapper for HTML Purifier for quick use.
@@ -8,12 +10,15 @@
 
 /**
  * Purify HTML.
- * @param string $html String HTML to purify
- * @param mixed $config Configuration to use, can be any value accepted by
- *        HTMLPurifier_Config::create()
+ *
+ * @param string $html   String HTML to purify
+ * @param mixed  $config Configuration to use, can be any value accepted by
+ *                       HTMLPurifier_Config::create()
+ *
  * @return string
+ * @throws HTMLPurifier_Exception
  */
-function HTMLPurifier($html, $config = null)
+function HTMLPurifier($html, $config = null): string
 {
     static $purifier = false;
     if (!$purifier) {
@@ -21,5 +26,3 @@ function HTMLPurifier($html, $config = null)
     }
     return $purifier->purify($html, $config);
 }
-
-// vim: et sw=4 sts=4
