@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @file
  * This file was auto-generated by generate-includes.php and includes all of
@@ -10,7 +12,7 @@
  * Changes to include_path are not necessary.
  */
 
-$__dir = dirname(__FILE__);
+$__dir = __DIR__;
 
 require_once $__dir . '/HTMLPurifier.php';
 require_once $__dir . '/HTMLPurifier/Arborize.php';
