@@ -1,14 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Super-class for definition datatype objects, implements serialization
  * functions for the class.
  */
 abstract class HTMLPurifier_Definition
 {
-
     /**
      * Has setup() been called yet?
+     *
      * @type bool
      */
     public $setup = false;
@@ -21,12 +23,14 @@ abstract class HTMLPurifier_Definition
      * is used and any writes to the raw definition object are short
      * circuited.  See enduser-customize.html for the high-level
      * picture.
+     *
      * @type bool
      */
     public $optimized = null;
 
     /**
      * What type of definition is it?
+     *
      * @type string
      */
     public $type;
@@ -34,22 +38,23 @@ abstract class HTMLPurifier_Definition
     /**
      * Sets up the definition object into the final form, something
      * not done by the constructor
+     *
      * @param HTMLPurifier_Config $config
      */
-    abstract protected function doSetup($config);
+    abstract protected function doSetup(HTMLPurifier_Config $config);
 
     /**
      * Setup function that aborts if already setup
+     *
      * @param HTMLPurifier_Config $config
      */
-    public function setup($config)
+    public function setup(HTMLPurifier_Config $config)
     {
         if ($this->setup) {
             return;
         }
+
         $this->setup = true;
         $this->doSetup($config);
     }
 }
-
-// vim: et sw=4 sts=4
