@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Global exception class for HTML Purifier; any exceptions we throw
  * are from here.
@@ -8,5 +10,3 @@ class HTMLPurifier_Exception extends Exception
 {
 
 }
-
-// vim: et sw=4 sts=4
