@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 /**
  * Exception type for HTMLPurifier_VarParser
@@ -7,5 +8,3 @@ class HTMLPurifier_VarParserException extends HTMLPurifier_Exception
 {
 
 }
-
-// vim: et sw=4 sts=4
